refactor: cosmos db interface
