+++ packages/core/src/bindings/interfaces/triggers/index.ts
@@ -1,3 +1,4 @@
 export * from './http-trigger.binding';
 export * from './timer-trigger.binding';
 export * from './cosmos-db-trigger.binding';
+export * from './cosmos-db.binding';
